# By Hoa V. Dinh (29) and others
# Via Hoà V. DINH (21) and others
* 'master' of https://github.com/dinhviethoa/libetpan: (111 commits)
  Update parameter for mailimap_has_extension and mailimap_has_authentication to use const char *
  Fixes of the workaround for qq.com IMAP server
  Implemented workaround for qq mail (#200)
  Fixed #198, #173: parse missing closing quote if it's not parsed as part of a word with an opening quote in a phrase
  Update mailimap.c
  Update mailimap.h
  Update mailimap.h
  Add macro ENABLE_163_WORKAROUND
  Send top-level multiple search without parenthesis
  Use closesocket() on windows
  Update mailimap.c
  Update mailimap.c
  Update mailimap.h
  Update mailimap_types.h
  Update mailimap.h
  Update mailimap.c
  Update mailimap.c
  Update mailimap.c
  Update mailimap.h
  Update mailimap.h
  ...

Conflicts:
	build-mac/libetpan.xcodeproj/project.pbxproj
	build-mac/libetpan.xcodeproj/xcshareddata/xcschemes/libetpan ios.xcscheme
	build-mac/libetpan.xcodeproj/xcshareddata/xcschemes/libetpan.xcscheme
	build-mac/libetpan.xcodeproj/xcshareddata/xcschemes/static libetpan.xcscheme

+++ src/data-types/mailstream_cancel.c
@@ -1,3 +1,33 @@
+/*
+ * libEtPan! -- a mail stuff library
+ *
+ * Copyright (C) 2001, 2014 - DINH Viet Hoa
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the libEtPan! project nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHORS AND CONTRIBUTORS ``AS IS'' AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHORS OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+ * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
+ */
 #include "mailstream_cancel.h"
 
 #ifdef HAVE_CONFIG_H
@@ -144,7 +174,7 @@ void mailstream_cancel_notify(struct mailstream_cancel * cancel)
   
   ch = 0;
 #ifndef WIN32
-  r = write(cancel->ms_fds[1], &ch, 1);
+  r = (int) write(cancel->ms_fds[1], &ch, 1);
 #else
   SetEvent(ms_internal->event);
 #endif
@@ -155,7 +185,7 @@ void mailstream_cancel_ack(struct mailstream_cancel * cancel)
 #ifndef WIN32
   char ch;
   int r;
-  r = read(cancel->ms_fds[0], &ch, 1);
+  r = (int) read(cancel->ms_fds[0], &ch, 1);
 #endif
 }
